refactor: move NoResultMessage to components dir
