+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2019-02-21T12:53:08.493Z\n"
-"PO-Revision-Date: 2019-02-21T12:53:08.493Z\n"
+"POT-Creation-Date: 2019-02-21T13:14:18.513Z\n"
+"PO-Revision-Date: 2019-02-21T13:14:18.514Z\n"
 
 msgid "Cancel"
 msgstr ""
