+++ src/components/NoResultsMessage.js
@@ -8,7 +8,7 @@ const NoResultsMessage = ({ additionalStyles }) => (
         id="no-std-report-find-message-id"
         style={{ textAlign: 'center', ...additionalStyles }}
     >
-        {i18n.t(i18nKeys.messages.noResultsFound)}
+        {i18n.t('No results have been found')}
     </p>
 )
