+++ src/pages/StandardReport.js
@@ -8,7 +8,7 @@ import manageError from '../utils/pageEnhancers/manageError.HOC'
 import { Snackbar } from '../components/feedback/Snackbar'
 import { SectionHeadline } from '../components/SectionHeadline'
 import SearchBox from './standard-report/SearchBox'
-import NoResultsMessage from './standard-report/NoResultsMessage'
+import NoResultsMessage from '../components/NoResultsMessage'
 import AddReportButton from './standard-report/AddReportButton'
 import ActionComponent from './standard-report/ActionComponent'
 import StyledHtmlReport from './standard-report/StyledHtmlReport'
